Refactor CLI commands (#1075)
* Refactor CLI commands

* Update opsdroid/__main__.py

* Update opsdroid/__main__.py

* Linting and tests

* Add deprecation warnings and test

* Refactor cli code out into opsdroid.cli module

* Only warn on old syntax

* Docs and docker

* Update opsdroid/cli/utils.py
