+++ tests/test_connector.py
@@ -4,11 +4,11 @@ import asyncio
 import asynctest
 import asynctest.mock as amock
 
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 from opsdroid.core import OpsDroid
 from opsdroid.connector import Connector, register_event
 from opsdroid.events import Event, Message, Reaction
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 
 
 class TestConnectorBaseClass(unittest.TestCase):
