+++ tests/test_connector_github.py
@@ -8,7 +8,7 @@ import unittest.mock as mock
 import asynctest
 import asynctest.mock as amock
 
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 from opsdroid.core import OpsDroid
 from opsdroid.connector.github import ConnectorGitHub
 from opsdroid.events import Message
