+++ tests/test_connector_gitter.py
@@ -12,7 +12,7 @@ from aiohttp.client_reqrep import ClientResponse, RequestInfo
 from yarl import URL
 
 
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 from opsdroid.core import OpsDroid
 from opsdroid.connector.gitter import ConnectorGitter
 from opsdroid.events import Message
