+++ tests/test_connector_slack.py
@@ -11,7 +11,7 @@ from opsdroid.core import OpsDroid
 from opsdroid.connector.slack import ConnectorSlack
 from opsdroid.connector.slack.events import Blocks
 from opsdroid.events import Message, Reaction
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 
 
 class TestConnectorSlack(unittest.TestCase):
