+++ tests/test_message.py
@@ -3,7 +3,7 @@ import asynctest
 from opsdroid.core import OpsDroid
 from opsdroid.message import Message
 from opsdroid.connector import Connector
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 
 
 class TestMessage(asynctest.TestCase):
