+++ tests/test_parser_dialogflow.py
@@ -4,7 +4,7 @@ import asynctest.mock as amock
 
 from aiohttp import ClientOSError
 
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 from opsdroid.core import OpsDroid
 from opsdroid.matchers import match_dialogflow_action
 from opsdroid.events import Message
