+++ tests/test_parser_event_type.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock
 import asynctest
 import asynctest.mock as amock
 
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 from opsdroid.core import OpsDroid
 from opsdroid.matchers import match_event
 from opsdroid import events
