+++ tests/test_parser_luisai.py
@@ -4,7 +4,7 @@ import asynctest.mock as amock
 
 from aiohttp import ClientOSError
 
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 from opsdroid.core import OpsDroid
 from opsdroid.matchers import match_luisai_intent
 from opsdroid.events import Message
