+++ tests/test_parser_witai.py
@@ -4,7 +4,7 @@ import asynctest.mock as amock
 
 from aiohttp import ClientOSError
 
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 from opsdroid.core import OpsDroid
 from opsdroid.matchers import match_witai
 from opsdroid.events import Message
