+++ tests/test_web.py
@@ -3,7 +3,7 @@ import ssl
 import asynctest
 import asynctest.mock as amock
 
-from opsdroid.__main__ import configure_lang
+from opsdroid.cli.start import configure_lang
 from opsdroid.core import OpsDroid
 from opsdroid import web
 import aiohttp.web
